Move parameter display to its own tab
