+++ translations/ui-developer/en.json
@@ -73,10 +73,11 @@
   "permissionsInspector": "Permissions inspector",
   "permissionsInspector.fromModule": "From module <code>{name}</code>",
   "okapiConsole": "Okapi console",
-  "okapiConsole.url": "URL",
-  "okapiConsole.tenant": "Tenant",
-  "okapiConsole.user": "User",
-  "okapiConsole.locale": "Locale",
+  "okapiConsole.parameters": "Parameters",
+  "okapiConsole.parameters.url": "URL",
+  "okapiConsole.parameters.tenant": "Tenant",
+  "okapiConsole.parameters.user": "User",
+  "okapiConsole.parameters.locale": "Locale",
 
   "plugin-surface": "Plugin surface",
   "plugin-surface.type": "Plugin type:",
